# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_search_image_api_v1_search_image_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "filters": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filters"
          },
          "top_k": {
            "default": 10,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_search_image_api_v1_search_image_post",
        "type": "object"
      },
...
      "CandidateResponse": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "cfr_summary": {
            "$ref": "#/components/schemas/CFRSummary"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "figure_id": {
            "format": "uuid",
            "title": "Figure Id",
            "type": "string"
          },
          "image_ref": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "matched_regions": {
            "items": {
              "$ref": "#/components/schemas/RegionResponse"
            },
            "title": "Matched Regions",
            "type": "array"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "score_components": {
            "$ref": "#/components/schemas/ScoreComponents"
          },
          "search_id": {
            "format": "uuid",
            "title": "Search Id",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/SourceSummary"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "uncertainty": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Uncertainty",
            "type": "object"
          },
          "verification_state": {
            "$ref": "#/components/schemas/VerificationState"
          }
        },
        "required": [
          "candidate_id",
          "figure_id",
          "source",
          "score",
          "score_components",
          "cfr_summary",
          "verification_state",
          "search_id",
          "created_at"
        ],
        "title": "CandidateResponse",
        "type": "object"
      },
...
      "EditionSummary": {
        "properties": {
          "allow_redistribution": {
            "title": "Allow Redistribution",
            "type": "boolean"
          },
          "book_id": {
            "format": "uuid",
            "title": "Book Id",
            "type": "string"
          },
          "edition_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Note"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "license_status": {
            "title": "License Status",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          }
        },
        "required": [
          "id",
          "book_id",
          "name",
          "license_status",
          "allow_redistribution"
        ],
        "title": "EditionSummary",
        "type": "object"
      },
...
      "FigureResponse": {
        "properties": {
          "assertions": {
            "items": {
              "$ref": "#/components/schemas/FunctionalAssertionResponse"
            },
            "title": "Assertions",
            "type": "array"
          },
          "asset": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssetRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "bbox": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bbox"
          },
          "data_status": {
            "$ref": "#/components/schemas/DataStatus"
          },
          "evidences": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Evidences",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionResponse"
            },
            "title": "Regions",
            "type": "array"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationResponse"
            },
            "title": "Relations",
            "type": "array"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "text_chunks": {
            "items": {
              "$ref": "#/components/schemas/TextChunkResponse"
            },
            "title": "Text Chunks",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "page_id"
        ],
        "title": "FigureResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestionJobRequest": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "manifest_name": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+\\.json$",
            "title": "Manifest Name",
            "type": "string"
          }
        },
        "required": [
          "manifest_name"
        ],
        "title": "IngestionJobRequest",
        "type": "object"
      },
      "IngestionJobResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dry_run": {
            "title": "Dry Run",
            "type": "boolean"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "manifest_name": {
            "title": "Manifest Name",
            "type": "string"
          },
          "stats": {
            "additionalProperties": true,
            "title": "Stats",
            "type": "object"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "manifest_name",
          "status",
          "dry_run",
          "created_at",
          "updated_at"
        ],
        "title": "IngestionJobResponse",
        "type": "object"
      },
...
      "TextSearchRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "query": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "TextSearchRequest",
        "type": "object"
      },
...
      "VerificationRequest": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "state": {
            "$ref": "#/components/schemas/VerificationState"
          }
        },
        "required": [
          "state"
        ],
        "title": "VerificationRequest",
        "type": "object"
      },
      "VerificationResponse": {
        "properties": {
          "candidate_id": {
            "format": "uuid",
            "title": "Candidate Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "state": {
            "$ref": "#/components/schemas/VerificationState"
          }
        },
        "required": [
          "id",
          "candidate_id",
          "state",
          "created_at"
        ],
        "title": "VerificationResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "面向传统技术图像的证据感知检索 V1 API（离线工程基线）",
    "title": "机图索隐后端",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets/{asset_id}": {
      "get": {
        "operationId": "asset_detail_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Asset Detail"
      }
    },
    "/api/v1/associations/{candidate_id}": {
      "get": {
        "operationId": "association_detail_api_v1_associations__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Association Detail"
      }
    },
    "/api/v1/associations/{candidate_id}/verify": {
      "post": {
        "operationId": "verify_association_api_v1_associations__candidate_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Association"
      }
    },
...
    "/api/v1/books/{book_id}/editions": {
      "get": {
        "operationId": "editions_api_v1_books__book_id__editions_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Book Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EditionSummary"
                  },
                  "title": "Response Editions Api V1 Books  Book Id  Editions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Editions"
      }
    },
...
    "/api/v1/figures/{figure_id}": {
      "get": {
        "operationId": "figure_detail_api_v1_figures__figure_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "figure_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Figure Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FigureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Figure Detail"
      }
    },
...
    "/api/v1/ingestion/jobs": {
      "post": {
        "operationId": "create_ingestion_job_api_v1_ingestion_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ingestion Job"
      }
    },
    "/api/v1/ingestion/jobs/{job_id}": {
      "get": {
        "operationId": "ingestion_job_detail_api_v1_ingestion_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingestion Job Detail"
      }
    },
    "/api/v1/pages/{page_id}": {
      "get": {
        "operationId": "page_detail_api_v1_pages__page_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Page Detail"
      }
    },
    "/api/v1/search/image": {
      "post": {
        "operationId": "search_image_api_v1_search_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_search_image_api_v1_search_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Image"
      }
    },
    "/api/v1/search/region": {
      "post": {
        "operationId": "search_region_api_v1_search_region_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegionSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Region"
      }
    },
    "/api/v1/search/text": {
      "post": {
        "operationId": "search_text_api_v1_search_text_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Text"
      }
    },
    "/api/v1/search/{search_id}": {
      "get": {
        "operationId": "search_detail_api_v1_search__search_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "search_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Search Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Detail"
      }
    }
  }
}